Renamed the dataset subpackage to datadoc
