+++ docs/api_reference/datadoc/clitool.md
@@ -1,0 +1,3 @@
+# clitool
+
+::: tripper.datadoc.clitool
