+++ docs/api_reference/datadoc/dataaccess.md
@@ -1,0 +1,3 @@
+# dataaccess
+
+::: tripper.datadoc.dataaccess
