+++ docs/api_reference/datadoc/dataset.md
@@ -1,0 +1,3 @@
+# dataset
+
+::: tripper.datadoc.dataset
