+++ docs/api_reference/datadoc/tabledoc.md
@@ -1,0 +1,3 @@
+# tabledoc
+
+::: tripper.datadoc.tabledoc
