--- docs/api_reference/dataset/dataaccess.md
@@ -1,3 +1,0 @@
-# dataaccess
-
-::: tripper.dataset.dataaccess
