--- docs/api_reference/dataset/datadoc.md
@@ -1,3 +1,0 @@
-# datadoc
-
-::: tripper.dataset.datadoc
