--- docs/api_reference/dataset/dataset.md
@@ -1,3 +1,0 @@
-# dataset
-
-::: tripper.dataset.dataset
