--- docs/api_reference/dataset/tabledoc.md
@@ -1,3 +1,0 @@
-# tabledoc
-
-::: tripper.dataset.tabledoc
