+++ docs/datadoc/introduction.md
@@ -9,10 +9,10 @@ The data documentation is based on small [JSON-LD documents], each documenting a
 Examples of resources can be a dataset, an instrument, a sample, etc.
 All resources are uniquely identified by their IRI.
 
-The primary focus of the [tripper.dataset] module is to document datasets such that they are consistent with the [DCAT vocabulary], but at the same time easily extended additional semantic meaning provided by other ontologies.
+The primary focus of the [tripper.datadoc] module is to document datasets such that they are consistent with the [DCAT vocabulary], but at the same time easily extended additional semantic meaning provided by other ontologies.
 It is also easy to add and relate the datasets to other types of documents, like people, instruments and samples.
 
-The [tripper.dataset] module provides a Python API for documenting resources at all four levels of data documentation, including:
+The [tripper.datadoc] module provides a Python API for documenting resources at all four levels of data documentation, including:
 
 - **Cataloguing**: Storing and accessing *documents* based on their IRI and data properties.
   (Addressed FAIR aspects: *findability* and *accessibility*).
@@ -30,7 +30,7 @@ The figure below shows illustrates how a dataset is documented in a triplestore.
 
 Resource types
 --------------
-The [tripper.dataset] module include the following set of predefined resource types:
+The [tripper.datadoc] module include the following set of predefined resource types:
 
 - **dataset**: Individual of [dcat:Dataset] and [emmo:DataSet].
 - **distribution**: Individual of [dcat:Distribution].
@@ -43,7 +43,7 @@ Future releases will support adding custom resource types.
 
 
 
-[tripper.dataset]: https://emmc-asbl.github.io/tripper/latest/api_reference/dataset/dataset
+[tripper.datadoc]: https://emmc-asbl.github.io/tripper/latest/datadoc/introduction
 [DCAT vocabulary]: https://www.w3.org/TR/vocab-dcat-3/
 [DLite]: https://github.com/SINTEF/dlite
 [YAML]: https://yaml.org/
@@ -58,10 +58,10 @@ Future releases will support adding custom resource types.
 [emmo:DataSet]: https://w3id.org/emmo#EMMO_194e367c_9783_4bf5_96d0_9ad597d48d9a
 [oteio:Generator]: https://w3id.org/emmo/domain/oteio/Generator
 [oteio:Parser]: https://w3id.org/emmo/domain/oteio/Parser
-[save_dict()]: ../../api_reference/dataset/dataset/#tripper.dataset.dataset.save_dict
-[as_jsonld()]: ../../api_reference/dataset/dataset/#tripper.dataset.dataset.as_jsonld
+[save_dict()]: ../../api_reference/datadoc/dataset/#tripper.datadoc.dataset.save_dict
+[as_jsonld()]: ../../api_reference/datadoc/dataset/#tripper.datadoc.dataset.as_jsonld
 [save_datadoc()]:
-../../api_reference/dataset/dataset/#tripper.dataset.dataset.save_datadoc
+../../api_reference/datadoc/dataset/#tripper.datadoc.dataset.save_datadoc
 [semdata.yaml]: https://raw.githubusercontent.com/EMMC-ASBL/tripper/refs/heads/master/tests/input/semdata.yaml
-[semdata.csv]: https://raw.githubusercontent.com/EMMC-ASBL/tripper/refs/heads/tabledoc-csv/tests/input/semdata.csv
-[TableDoc]: https://emmc-asbl.github.io/tripper/latest/api_reference/dataset/dataset/#tripper.dataset.tabledoc.TableDoc
+[semdata.csv]: https://raw.githubusercontent.com/EMMC-ASBL/tripper/refs/heads/master/tests/input/semdata.csv
+[TableDoc]: https://emmc-asbl.github.io/tripper/latest/api_reference/datadoc/tabledoc/#tripper.datadoc.tabledoc.TableDoc
