+++ tripper/datadoc/dataset.py
@@ -27,7 +27,7 @@ Functions for interaction with OTEAPI:
 ---
 
 [DCAT]: https://www.w3.org/TR/vocab-dcat-3/
-[JSON-LD context]: https://raw.githubusercontent.com/EMMC-ASBL/tripper/refs/heads/dataset/tripper/context/0.2/context.json
+[JSON-LD context]: https://raw.githubusercontent.com/EMMC-ASBL/tripper/refs/heads/master/tripper/context/0.2/context.json
 
 """
 
@@ -887,7 +887,7 @@ def search_iris(ts: Triplestore, type=None, **kwargs) -> "List[str]":
             )
 
     SeeAlso:
-    [resource type]: https://emmc-asbl.github.io/tripper/latest/dataset/introduction/#resource-types
+    [resource type]: https://emmc-asbl.github.io/tripper/latest/datadoc/introduction/#resource-types
     """
     # Special handling of @id
     id = kwargs.pop("@id") if "@id" in kwargs else kwargs.pop("_id", None)
